establish primary run file/main event loop
