+++ config.py
@@ -1,0 +1,1 @@
+DISCORD_BOT_TOKEN: str = "MTAxMjg0NTY3NjcwNDU4Mzc5MA.GO7xyd.mWCAyxmKELGSa6hr6xKOQ4tDARsvfz2lRNdBKA"
